expert(professional) 권한 해결 version(int ->Long) 변경

+++ src/main/java/com/kaidey/yakchatproject/domain/onboarding/dto/ExpertOnboardingRes.java
@@ -8,5 +8,5 @@ import java.util.List;
 public record ExpertOnboardingRes(
         Long userId,
         boolean completed,
-        Integer profileVersion
+        Long profileVersion
 ) {}
